A lot of cleanup & restructurising
