+++ packages/client/src/contexts/CategoryMappingContext.ts
@@ -1,0 +1,11 @@
+import React from "react";
+import { preprocessCategoryData } from "../core/tools";
+import { CategoryMapping } from "../types";
+import categories from '../data/categories.json';
+
+export const categoryMapping = preprocessCategoryData(categories);
+
+const CategoryMappingContext = React.createContext<CategoryMapping>(categoryMapping);
+
+export default CategoryMappingContext;
+
