+++ packages/client/src/index.tsx
@@ -2,16 +2,14 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 // import reportWebVitals from './reportWebVitals';
-import { RouterProvider } from 'react-router-dom';
-
-import appRouter from './routing';
+import App from './App';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
 );
 root.render(
   <React.StrictMode>
-    <RouterProvider router={appRouter} />
+    <App />
   </React.StrictMode>
 );
