+++ packages/client/src/routes/LoginPage.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export default function LoginPage(): React.JSX.Element {
+  return (
+    <div>
+      <h1>Login page</h1>
+      <p><a href="/">Take me back to home page</a></p>
+    </div>
+  );
+}
+
